AU-1872 Make own submodule for audit logging
